modificações na ordem de execução realizada
