+++ src/com/ufg/i4soft/droidrepair/controller/ManagePlugin.java
@@ -12,47 +12,23 @@ import java.util.Optional;
 
 public class ManagePlugin {
 
-    public void start(){
+    public void start() {
 
-        Optional<String> pathOptional = Optional.ofNullable(collectPath());
+        ChooseRepair.showTypeRepair();
 
-        boolean file_valid;
+        ProjectData data = new ProjectData();
+        data.setTypeRepair(ChooseRepair.getTyperepair());
 
-        FilterData filterData = new FilterData();
-
-        try {
-
-            file_valid = filterData.verifyExtensionFile(pathOptional.get());
-
-            if (file_valid) {
-
-                ChooseRepair.showTypeRepair();
-
-                ProjectData data = new ProjectData();
-                data.setTypeRepair(ChooseRepair.getTyperepair());
-
-                selectRepair(data.getTypeRepair(), pathOptional.get());
-            }
-
-        } catch (NoSuchElementException e){
-
-            Messages.showMessageDialog("Arquivo para reparo ausente", "Arquivo Ausente", Messages.getInformationIcon());
-        }
+        selectRepair(data.getTypeRepair());
     }
 
-    private String collectPath() {
-
-        MainWindows windows = new MainWindows();
-        return windows.viewChooseFile("Selecione O Diretório Do Arquivo", "Selecione o arquivo que deseja submeter à execução do angelix");
-    }
-
-    private void selectRepair(String repair, String path) {
+    private void selectRepair(String repair) {
 
         switch (repair) {
 
             case "angelix":
                 ManageAngelix angelix = new ManageAngelix();
-                angelix.startRepairTool(path);
+                angelix.startRepairTool();
                 break;
 
             case "genprog":
